Migrate from Netlify to Cloudflare Pages
To anyone wondering why this happened - please start by reading
https://news.ycombinator.com/item?id=39520776

The company issued a half-hearted apology on Twitter and admitted
little, if any fault. They also made no specific commitment whatsoever
to prevent the same thing from happening again.

As a result, I filed a support ticket to ask for more information on
what guarantees free-plan users have that they will not be invoiced
hundreds of thousands of dollars in the future.

The initial reply I got was an incredibly patronizing suggestion that
I should talk to their AI chatbot since they are too busy for me. I
pointed out that this was not an appropriate response, and
subsequently received a response from a human saying that no
additional information would be released.

I informed Netlify that they had one month to provide some concrete
evidence (of any sort) that they had made progress to prevent more
fraudulent charges. When I followed up a month later, I was informed
that no information would be provided, and furthermore "we won't
answer any questions related to this issue in this ticket anymore. We
will be closing this ticket and we won't provide any further responses
here."

So, fuck those guys. It would be insane to run any application on
infrastructure owned by a company that thinks it is totally peachy to
send $100k+ invoices to free-plan users, and has absolutely no plan to
change any of their practices.