Vercel => Cloudflare
If you feel safe hosting anything on Cloudflare, read this:
https://robindev.substack.com/p/cloudflare-took-down-our-website

Then take note of the fact that not only has Cloudflare not
apologized, or refunded or restored anything, or admitted fault in any
way, but the CEO publicly dumped on the person who posted it.

At least when other platforms screw up this badly, they apologize and
refund. I even reached out to Cloudflare myself for a response, and
they told me curtly to go fuck myself (more or less).